requirements.txt
autogen>=0.9.0
python-wappalyzer==0.3.1
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2